Extract and Predict stock prices from yfinance
It's a simple and easy to understand model using Monte Carlo Simulation technique.
Monte Carlo simulation is a computational method that uses random sampling and statistical analysis to model and solve complex problems. It is named after the famous casino city of Monte Carlo, Monaco, because the method relies on generating a large number of random samples, similar to the random outcomes in games of chance.